# api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SpecGen API",
    "version": "1.0.0",
    "description": "API documentation for SpecGen - a speculative fiction generator",
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/database/download": {
      "get": {
        "summary": "Download the database file",
        "description": "Retrieve the current database file in JSON format.",
        "responses": {
          "200": {
            "description": "Database data retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "No database data found."
          },
          "500": {
            "description": "Failed to download database."
          }
        }
      }
    },
    "/api/database/generations/download": {
      "get": {
        "summary": "Download the generations database",
        "description": "Retrieve all generated content from the SQLite database in JSON format.",
        "responses": {
          "200": {
            "description": "Generations data retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "No generations data found."
          },
          "500": {
            "description": "Failed to download generations database."
          }
        }
      }
    },
    "/api/database/restore": {
      "post": {
        "summary": "Restore the database from a file",
        "description": "Restore the database using an uploaded JSON file. The file should contain valid \"categories\" and \"parameters\" arrays.",
        "consumes": [
          "multipart/form-data"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Database restored successfully."
          },
          "400": {
            "description": "No file uploaded, invalid JSON file, or invalid database structure."
          },
          "500": {
            "description": "Failed to restore database."
          }
        }
      }
    },
    "/api/database/generations/restore": {
      "post": {
        "summary": "Restore the generations database from a file",
        "description": "Restore the generations database using an uploaded JSON file. The file should contain a valid \"generations\" array.",
        "consumes": [
          "multipart/form-data"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generations database restored successfully."
          },
          "400": {
            "description": "No file uploaded, invalid JSON file, or invalid database structure."
          },
          "500": {
            "description": "Failed to restore generations database."
          }
        }
      }
    },
    "/api/database/reset": {
      "post": {
        "summary": "Reset the database to an empty state",
        "description": "Clear all data from the database by resetting it to an empty state.",
        "responses": {
          "200": {
            "description": "Database reset successfully."
          },
          "500": {
            "description": "Failed to reset database."
          }
        }
      }
    },
    "/api/database/generations/reset": {
      "post": {
        "summary": "Reset the generations database to an empty state",
        "description": "Clear all data from the generations database by resetting it to an empty state.",
        "responses": {
          "200": {
            "description": "Generations database reset successfully."
          },
          "500": {
            "description": "Failed to reset generations database."
          }
        }
      }
    },
    "/api/database/reset-all": {
      "post": {
        "summary": "Reset all databases to an empty state",
        "description": "Clear all data from both the primary database and generations database.",
        "responses": {
          "200": {
            "description": "All databases reset successfully."
          },
          "500": {
            "description": "Failed to reset databases."
          }
        }
      }
    },
...
  },
...
}